setTrialPeriod fucntion correction
corrected this in instance where we are using planCombination instead of plan to fetch the actual plan to use its trial date and trial interval. Since those are not found in PlanCombination trial end date was being set as null in case a PlanCombination was being passed upon calling newSubscription function.